+++ HISTORY.md
@@ -46,6 +46,7 @@
 ### 2019-03-04
 * GUI: deprecated storing projects in localStorage
 * GUI: reenabled publish / unpublish buttons in the project dialog
+* GUI: spread project dialog action buttons over 2 rows
 
 ### 2019-02-26
 * Symbols: new "globe" symbol